Add function to src/card
