+++ test/Card-test.js
@@ -1,7 +1,7 @@
 const chai = require('chai');
 const expect = chai.expect;
 
-const { createCard, checkForWin } = require('../src/card');
+const { createCard } = require('../src/card');
 
 describe('card', function() {
   it('should be a function', function() {
@@ -16,44 +16,4 @@ describe('card', function() {
     expect(card.answers).to.deep.equal(['object', 'array', 'function']);
     expect(card.correctAnswer).to.equal('object');
   });  
-});
-
-
-describe('checkForWin', function() {
-// happy
-  it('should return correct! when answer is correct', function() {
-    // 1. Setup
-      // Create mock player
-      // create variable to hold correct word
-    const player = {
-      name: 'Mike',
-      winningWord: []
-    };
-    const correctWord = 'object';
-    // 2. Execution
-      // Run appropriate functions that execute the behavior indicated by test title
-      // Pass in the arguments needed to test the specific scenario trying to test
-    checkForWin('object', correctWord);
-    // Assertion
-      // make an assertion to verify that after executing certain function, we end up with what we expect
-    expect(player.winningWord).to.deep.equal(['object']);
-  });
-// sad 
-  it('should return incorrect! when answer is incorrect', function() {
-    // 1. Setup
-    // create mock player
-    // create variable to hold correct word
-    const player = {
-      name: 'Mike',
-      winningWord: []
-    };
-    const correctWord = 'object';
-    // 2. Execution
-    // Run appropriate functions that execute the behavior indicated by test title
-    // Pass in the arguments needed to test the specific scenario trying to test
-    checkForWin('array', correctWord);
-    // 3. Assertion
-      // make an assertion to verify that after executing certain function, we end up with what we expect
-      expect(player.winningWord).to.deep.equal([]);
-  });
 });
